+++ src/main/java/com/structurizr/dsl/StructurizrDslParser.java
@@ -26,6 +26,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
     private static final Pattern COMMENT_PATTERN = Pattern.compile("^\\s*?(//|#).*$");
     private static final String MULTI_LINE_COMMENT_START_TOKEN = "/*";
     private static final String MULTI_LINE_COMMENT_END_TOKEN = "*/";
+    private static final String MULTI_LINE_SEPARATOR = "\\";
 
     private static final Pattern STRING_SUBSTITUTION_PATTERN = Pattern.compile("(\\$\\{[a-zA-Z0-9-_.]+?})");
 
@@ -152,10 +153,44 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
         parse(lines, new File("."));
     }
 
-    void parse(List<String> lines, File dslFile) throws StructurizrDslParserException {
+    private List<DslLine> preProcessLines(List<String> lines) {
+        List<DslLine> dslLines = new ArrayList<>();
+
         int lineNumber = 1;
+        StringBuilder buf = new StringBuilder();
+        boolean lineComplete = true;
+
         for (String line : lines) {
+            if (line.endsWith(MULTI_LINE_SEPARATOR)) {
+                buf.append(line, 0, line.length()-1);
+                lineComplete = false;
+            } else {
+                if (lineComplete) {
+                    buf.append(line);
+                } else {
+                    buf.append(line.stripLeading());
+                    lineComplete = true;
+                }
+            }
+
+            if (lineComplete) {
+                dslLines.add(new DslLine(buf.toString(), lineNumber));
+                buf = new StringBuilder();
+            }
+
+            lineNumber++;
+        }
+
+        return dslLines;
+    }
+
+    void parse(List<String> lines, File dslFile) throws StructurizrDslParserException {
+        List<DslLine> dslLines = preProcessLines(lines);
+
+        for (DslLine dslLine : dslLines) {
             boolean includeInDslSourceLines = true;
+
+            String line = dslLine.getSource();
 
             if (line.startsWith(BOM)) {
                 // this caters for files encoded as "UTF-8 with BOM"
@@ -827,13 +862,11 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                 if (includeInDslSourceLines) {
                     dslSourceLines.add(line);
                 }
-
-                lineNumber++;
             } catch (Exception e) {
                 if (e.getMessage() != null) {
-                    throw new StructurizrDslParserException(e.getMessage(), lineNumber, line);
+                    throw new StructurizrDslParserException(e.getMessage(), dslLine.getLineNumber(), line);
                 } else {
-                    throw new StructurizrDslParserException(e.getClass().getSimpleName(), lineNumber, line);
+                    throw new StructurizrDslParserException(e.getClass().getSimpleName(), dslLine.getLineNumber(), line);
                 }
             }
         }